feat: add splitter function and its test

+++ equation.go
@@ -1,6 +1,6 @@
 package equation
 
-func Solve(expression string) int {
+func Solve(expression string) float64 {
 
 	return 2
 }
